requirements.txt
streamlit==1.32.0
langchain==0.1.12
openai==1.15.0
matplotlib==3.8.3
pandas==2.2.1
python-dotenv==1.0.1
numpy==1.26.4
yfinance==0.2.36 